+++ R/colormap-package.r
@@ -1,7 +1,6 @@
-#' A package to generate colors from a list of 44 pre-defined palettes.
+#' A package to generate colors from a list of 44 pre-defined palettes
 #'
 #' @name colormap
 #' @docType package
 #' @author Bhaskar V. Karambelkar
-#' @importFrom V8 new_context
 NULL