+++ NEWS.md
@@ -1,3 +1,9 @@
+# colormap 0.1.1
+
+* Added code to evenly distribute colors over the whole range of the palette.
+* Added Input Validation.
+* Added scale_fill_colormap and scale_color_colormap for ggplot2.
+
 # colormap 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.